requirements.txt
Django
djangorestframework
django-cors-headers
celery
drf-yasg
django-environ
mysqlclient
requests
pika